Últimos cambios del Input
