+++ FORGE/projects/Input/Input.vcxproj.filters
@@ -15,9 +15,6 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\src\Input\InputTest.cpp">
-      <Filter>Archivos de origen</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\Input\Input.cpp">
       <Filter>Archivos de origen</Filter>
     </ClCompile>
