+++ FORGE/src/Input/Input.h
@@ -4,20 +4,77 @@
 #include <unordered_map>
 #include <array>
 #include "SDL.h"
-#include "KeyNames.h"
 
 #define CONTROLLER_AXIS_MAX 32767.0f
 #define CONTROLLER_AXIS_MIN -32768.0f
 #define CONTROLLER_AXIS_POS_DEADZONE (CONTROLLER_AXIS_MAX * 0.3f)
 #define CONTROLLER_AXIS_NEG_DEADZONE (CONTROLLER_AXIS_MIN * 0.3f)
 
+const int KEYNAMES_SIZE = 62;
+
+/// <summary>
+/// Nombres de las teclas desde fuera del motor.
+/// Cada fila del teclado en una fila, si se quieren agregar teclas hacerlo por el final.
+/// </summary>
+enum KeyNames {
+	K_ESC, K_F1, K_F2, K_F3, K_F4, K_F5, K_F6, K_F7, K_F8, K_F9, K_F10, K_F11, K_F12,
+	K_1, K_2, K_3, K_4, K_5, K_6, K_7, K_8, K_9, K_0, K_RETURN,
+	K_TAB, K_Q, K_W, K_E, K_R, K_T, K_Y, K_U, K_I, K_O, K_P,
+	K_CAPS, K_A, K_S, K_D, K_F, K_G, K_H, K_J, K_K, K_L, K_ENTER,
+	K_SHIFT, K_Z, K_X, K_C, K_V, K_B, K_N, K_M,
+	K_CTRL, K_CONTEXT, K_ALT, K_SPACE, K_LEFT, K_UP, K_DOWN, K_RIGHT
+};
+
+/// <summary>
+/// Nombres de los botones del raton desde fuera del motor
+/// </summary>
+enum MouseNames {
+	M_LEFT, M_MIDDLE, M_RIGHT
+};
+
+/// <summary>
+/// Nombres de los botones del mando desde fuera del motor
+/// </summary>
+enum ControllerButtonNames {
+	C_BUTTON_INVALID = -1,
+	C_BUTTON_A, C_BUTTON_B, C_BUTTON_X, C_BUTTON_Y,
+	C_BUTTON_BACK, C_BUTTON_GUIDE, C_BUTTON_START,
+	C_BUTTON_LEFTSTICK, C_BUTTON_RIGHTSTICK,
+	C_BUTTON_LEFT_SHOULDER, C_BUTTON_RIGHT_SHOULDER,
+	C_BUTTON_DPAD_UP, C_BUTTON_DPAD_DOWN, C_BUTTON_DPAD_LEFT, C_BUTTON_DPAD_RIGHT,
+	C_BUTTON_MISC1,    /* Xbox Series X share button, PS5 microphone button, Nintendo Switch Pro capture button, Amazon Luna microphone button */
+	C_BUTTON_PADDLE1,  /* Xbox Elite paddle P1 (upper left, facing the back) */
+	C_BUTTON_PADDLE2,  /* Xbox Elite paddle P3 (upper right, facing the back) */
+	C_BUTTON_PADDLE3,  /* Xbox Elite paddle P2 (lower left, facing the back) */
+	C_BUTTON_PADDLE4,  /* Xbox Elite paddle P4 (lower right, facing the back) */
+	C_BUTTON_TOUCHPAD, /* PS4/PS5 touchpad button */
+	C_BUTTON_MAX
+};
+
+/// <summary>
+/// Nombres de los ejes del mando desde fuera del motor
+/// </summary>
+enum ControllerAxisNames {
+	C_AXIX_INVALID = -1,
+	C_AXIS_LEFTX, C_AXIS_LEFTY,
+	C_AXIS_RIGHTX, C_AXIS_RIGHTY,
+	C_AXIS_TRIGGERLEFT, C_AXIS_TRIGGERRIGHT,
+	C_AXIS_MAX
+};
+
 class Input {
 private:
 	static std::unique_ptr<Input> instance;
 
-	SDL_Event SDLPreviousEvent;
+	/// <summary>
+	/// Array dinamico que proporciona SDL con el estado de cada tecla. Si kS[key] es 1 entonces esta siendo pulsada,
+	/// de lo contrario sera 0
+	/// </summary>
 	const Uint8* keyboardState;
-	// true -> presionada/mantenida, false -> soltada
+	/// <summary>
+	/// Mapa con el estado previo de cada tecla. Si esta siendo presionada o mantenida entonces el valor es true, mientras
+	/// que si ha sido soltada el valor es false 
+	/// </summary>
 	std::unordered_map<KeyNames, bool> keys;
 
 	std::pair<int, int> mousePos;
@@ -37,8 +94,8 @@ private:
 	/// </summary>
 	const SDL_Scancode SCANCODE[KEYNAMES_SIZE] =
 	{
-		{SDL_SCANCODE_ESCAPE},{SDL_SCANCODE_F1}, {SDL_SCANCODE_F2}, {SDL_SCANCODE_F3}, {SDL_SCANCODE_F4},
-		{SDL_SCANCODE_F5},{SDL_SCANCODE_F6}, {SDL_SCANCODE_F7}, {SDL_SCANCODE_F8}, {SDL_SCANCODE_F9},
+		{SDL_SCANCODE_ESCAPE}, {SDL_SCANCODE_F1}, {SDL_SCANCODE_F2}, {SDL_SCANCODE_F3}, {SDL_SCANCODE_F4},
+		{SDL_SCANCODE_F5}, {SDL_SCANCODE_F6}, {SDL_SCANCODE_F7}, {SDL_SCANCODE_F8}, {SDL_SCANCODE_F9},
 		{SDL_SCANCODE_F10}, {SDL_SCANCODE_F11},	{SDL_SCANCODE_F12}, {SDL_SCANCODE_1}, {SDL_SCANCODE_2},
 		{SDL_SCANCODE_3}, {SDL_SCANCODE_4},	{SDL_SCANCODE_5},  {SDL_SCANCODE_6}, {SDL_SCANCODE_7},
 		{SDL_SCANCODE_8}, {SDL_SCANCODE_9},	{SDL_SCANCODE_0},  {SDL_SCANCODE_BACKSPACE}, {SDL_SCANCODE_TAB},
@@ -57,20 +114,20 @@ private:
 	/// <summary>
 	/// Guarda la posicion actual del raton
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
+	/// <param name="event">Evento al que reaccionar</param>
 	void onMouseMotion(const SDL_Event& event);
 
 	/// <summary>
 	/// Detecta si la rueda se ha movido hacia arriba o hacia abajo
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
+	/// <param name="event">Evento al que reaccionar</param>
 	void onMouseWheelMotion(const SDL_Event& event);
 
 	/// <summary>
 	/// Marca en el array de los botones del raton si este se pulsa o se deja de pulsar
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
-	/// <param name="down">- 'true' si esta siendo pulsado, 'false' si no</param>
+	/// <param name="event">Evento al que reaccionar</param>
+	/// <param name="down">Sera 'true' si esta siendo pulsado, 'false' si no</param>
 	void onMouseButton(const SDL_Event& event, bool down);
 
 	/// <summary>
@@ -110,19 +167,19 @@ private:
 	/// <summary>
 	/// Cambia el evento de Button Down a true
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
+	/// <param name="event">Evento al que reaccionar</param>
 	void onControllerButtonDown(const SDL_Event& event);
 
 	/// <summary>
 	/// Cambia el evento de Button UP a true
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
+	/// <param name="event">Evento al que reaccionar</param>
 	void onControllerButtonUp(const SDL_Event& event);
 
 	/// <summary>
 	/// Cambia el evento de Axis Motion a true
 	/// </summary>
-	/// <param name="event">- evento al que reaccionar</param>
+	/// <param name="event">Evento al que reaccionar</param>
 	void onControllerAxisMotion(const SDL_Event& event);
 
 	/// <summary>
@@ -161,19 +218,19 @@ public:
 	/// <summary>
 	/// Devuelve si se ha pulsado la tecla correspondiente
 	/// </summary>
-	/// <param name="k">- tecla</param>
+	/// <param name="k">Tecla a comprobar</param>
 	bool keyDown(KeyNames k);
 
 	/// <summary>
 	/// Devuelve si se esta manteniendo la tecla correspondiente
 	/// </summary>
-	/// <param name="k">- tecla</param>
+	/// <param name="k">Tecla a comprobar</param>
 	bool keyPressed(KeyNames k);
 
 	/// <summary>
 	/// Devuelve si se ha soltado la tecla correspondiente
 	/// </summary>
-	/// <param name="k">- tecla</param>
+	/// <param name="k">Tecla a comprobar</param>
 	bool keyUp(KeyNames k);
 
 	/// <summary>
@@ -201,25 +258,25 @@ public:
 	/// <summary>
 	/// Devuelve si se ha pulsado el boton del mando
 	/// </summary>
-	/// <param name="button">- indice del boton del mando</param>
+	/// <param name="button">Indice del boton del mando</param>
 	bool isControllerButtonDown(ControllerButtonNames button);
 
 	/// <summary>
 	/// Devuelve si se ha soltado el boton del mando
 	/// </summary>
-	/// <param name="button">- indice del boton del mando</param>
+	/// <param name="button">Indice del boton del mando</param>
 	bool isControllerButtonUp(ControllerButtonNames button);
 
 	/// <summary>
 	/// Devuelve el valor del eje del joystick del mando
 	/// </summary>
-	/// <param name="ax">- eje del mando</param>
+	/// <param name="ax">Eje del mando</param>
 	int getControllerAxis(ControllerAxisNames ax);
 
 	/// <summary>
 	/// Devuelve el valor normalizado del eje del joystick del mando
 	/// </summary>
-	/// <param name="ax">- eje del mando</param>
+	/// <param name="ax">Eje del mando</param>
 	float getNormalizedControllerAxis(ControllerAxisNames ax);
 
 	/// <summary>
